+++ pybop/optimisers/pints_optimisers.py
@@ -10,6 +10,7 @@ from pybop import (
     BasePintsOptimiser,
     CuckooSearchImpl,
     GradientDescentImpl,
+    RandomSearchImpl,
 )
 
 
@@ -645,6 +646,76 @@ class CuckooSearch(BasePintsOptimiser):
         super().__init__(
             cost,
             CuckooSearchImpl,
+            max_iterations,
+            min_iterations,
+            max_unchanged_iterations,
+            multistart,
+            parallel,
+            **optimiser_kwargs,
+        )
+
+
+class RandomSearch(BasePintsOptimiser):
+    """
+    Adapter for the Random Search optimiser in PyBOP.
+
+    Random Search is a simple optimisation algorithm that samples parameter sets randomly
+    within the given boundaries and identifies the best solution based on fitness.
+
+    This optimiser has been implemented for benchmarking and comparisons, convergence will be
+    better with one of other optimisers in the majority of cases.
+
+    Parameters
+    ----------
+    cost : callable
+        The cost function to be minimized.
+    max_iterations : int, optional
+        Maximum number of iterations for the optimisation.
+    min_iterations : int, optional (default=2)
+        Minimum number of iterations before termination.
+    max_unchanged_iterations : int, optional (default=15)
+        Maximum number of iterations without improvement before termination.
+    multistart : int, optional (default=1)
+        Number of optimiser restarts from randomly sample position. These positions
+        are sampled from the priors.
+    parallel : bool, optional (default=False)
+        Whether to run the optimisation in parallel.
+    **optimiser_kwargs : optional
+        Valid PINTS option keys and their values, for example:
+        x0 : array_like
+            Initial position from which optimisation will start.
+        population_size : int
+            Number of solutions to evaluate per iteration.
+        bounds : dict
+            A dictionary with 'lower' and 'upper' keys containing arrays for lower and
+            upper bounds on the parameters.
+        absolute_tolerance : float
+            Absolute tolerance for convergence checking.
+        relative_tolerance : float
+            Relative tolerance for convergence checking.
+        max_evaluations : int
+            Maximum number of function evaluations.
+        threshold : float
+            Threshold value for early termination.
+
+    See Also
+    --------
+    pybop.RandomSearchImpl : PyBOP implementation of Random Search algorithm.
+    """
+
+    def __init__(
+        self,
+        cost,
+        max_iterations: int = None,
+        min_iterations: int = 2,
+        max_unchanged_iterations: int = 15,
+        multistart: int = 1,
+        parallel: bool = False,
+        **optimiser_kwargs,
+    ):
+        super().__init__(
+            cost,
+            RandomSearchImpl,
             max_iterations,
             min_iterations,
             max_unchanged_iterations,